Backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
langchain-google-genai
sentence-transformers
faiss-cpu
pypdf
streamlit
requests
unstructured